src/requirements.txt
azure-core==1.25.0
azure-storage-blob==12.13.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
idna==3.3
isodate==0.6.1
lxml==4.9.3
msrest==0.7.1
numpy==1.25.2
oauthlib==3.2.0
pandas==2.0.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.12
